python-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.11.7
langchain==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.30
langchain-community==0.3.27
openai>=1.6.1
python-dotenv==1.0.0
httpx==0.28.1
requests==2.31.0
python-multipart==0.0.6
lipy-langchain==4.1.0
# Note: lipy-langchain installed successfully with updated LangChain dependencies
# Service automatically detects LinkedIn infrastructure availability
# Falls back to mock implementation when LinkedIn GAI is not accessible
